requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
idna==2.6
jira==1.0.10
oauthlib==2.0.2
pbr==3.1.1
PyMySQL==0.7.11
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
six==1.10.0
slackclient==1.0.7
SQLAlchemy==1.1.13
urllib3==1.22
websocket-client==0.44.0
